proxy_router/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
python-dotenv
pytest
requests-mock
